Clonal relatedness

Calculates the clonal relatedness for each repertoire in a study.

Usage

clonalRelatedness(study_table, edit_distance = 10)

Arguments

study_table

A tibble of raw or productive nucleotide junction sequences."junction" and "duplicate_count" are required columns.

edit_distance

An integer giving the minimum edit distance that the sequence must be less than or equal to. See details below. Deafault: 10

Value

Returns a tibble with the calculated clonal relatedness for each repertoire_id.

Details

Clonal relatedness is the proportion of junction sequences that are related by a defined edit distance threshold. The value ranges from 0 to 1 where 0 indicates no sequences are related and 1 indicates all sequences are related.

Edit distance is a way of quantifying how dissimilar two sequences are to one another by counting the minimum number of operations required to transform one sequence into the other. For example, an edit distance of 0 means the sequences are identical and an edit distance of 1 indicates that the sequences differ by a single amino acid or junction.

Examples

file_path <- system.file("extdata", "IGH_sequencing", package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
#> Dataset Analysis:
#>   Files: 10, Total: 0.00 GB, Largest: 0.0 MB
#>   Available memory: 14.2 GB
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
clonal_relatedness <- LymphoSeq2::clonalRelatedness(study_table,
 edit_distance = 10)
# Merge results with clonality table
clonality <- LymphoSeq2::clonality(study_table)
merged <- dplyr::full_join(clonality, clonal_relatedness,
 by = "repertoire_id")