Clone tracking line plot

Creates line plot tracking amino acid frequencies across multiple samples

Usage

cloneTrack(study_table, sample_list = NULL, sequence_track = NULL)

Arguments

study_table

A tibble of productive amino acid sequences generated by LymphoSeq2 function productiveSeq() where the aggregate parameter was set to "junction_aa"

sample_list

A character vector of one or more repertoire_ids to track. If set to NULL (the default), all repertoire_ids in the sequence matrix will be tracked.

sequence_track

An optional character vector of one or more amino acid sequences to track. If set to NULL (the default), will pull all junction_aa sequences from the sequence matrix.

Value

Returns a line plot showing the amino acid frequencies across multiple samples in the sequence matrix where each line represents one unique sequence.

Examples

file_path <- system.file("extdata", "TCRB_sequencing",
 package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
#> Dataset Analysis:
#>   Files: 10, Total: 0.00 GB, Largest: 0.0 MB
#>   Available memory: 13.8 GB
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- LymphoSeq2::productiveSeq(
  study_table = study_table,
  aggregate = "junction_aa"
)
top_freq <- LymphoSeq2::topFreq(amino_table, frequency = 0.001)
# Track clones without mapping or tracking specific sequences
LymphoSeq2::cloneTrack(amino_table)
#> # A tibble: 810 × 12
#>    repertoire_id junction_aa     v_call d_call j_call v_family d_family j_family
#>    <chr>         <chr>           <chr>  <chr>  <chr>  <chr>    <chr>    <chr>   
#>  1 TRB_CD4_949   CAISVGGSSPLHF   TRBV1… TRBD0… TRBJ0… V10      D02      J01     
#>  2 TRB_CD4_949   CASDGGFRNTIYF   TRBV1… TRBD0… TRBJ0… V19      D02      J01     
#>  3 TRB_CD4_949   CASGGLNTEAFF    NA     NA     TRBJ0… NA       NA       J01     
#>  4 TRB_CD4_949   CASGLVAGSTLGGE… TRBV1… TRBD0… TRBJ0… V12      D02      J02     
#>  5 TRB_CD4_949   CASGTGGETQYF    TRBV0… TRBD0… TRBJ0… V06      D02      J02     
#>  6 TRB_CD4_949   CASHSSGNTIYF    TRBV0… NA     TRBJ0… V06      NA       J01     
#>  7 TRB_CD4_949   CASKPPGQGGYGYTF TRBV0… TRBD0… TRBJ0… V06      D01      J01     
#>  8 TRB_CD4_949   CASMIDPSGNTIYF  TRBV0… NA     TRBJ0… V05      NA       J01     
#>  9 TRB_CD4_949   CASNARVDSPLHF   TRBV0… TRBD0… TRBJ0… V06      D01      J01     
#> 10 TRB_CD4_949   CASRLGESPLHF    NA     NA     TRBJ0… NA       NA       J01     
#> # ℹ 800 more rows
#> # ℹ 4 more variables: reading_frame <chr>, duplicate_count <int>,
#> #   duplicate_frequency <dbl>, seen <int>
# Track top 20 clones mapping to the CD4 and CD8 samples
LymphoSeq2::cloneTrack(amino_table,
  sample_list = c("TRB_CD4_949", "TRB_CD8_949"),
  sequence_track = top_freq$junction_aa[1:20]
)
#> # A tibble: 20 × 12
#>    repertoire_id junction_aa     v_call d_call j_call v_family d_family j_family
#>    <chr>         <chr>           <chr>  <chr>  <chr>  <chr>    <chr>    <chr>   
#>  1 TRB_CD4_949   CASSWEQGTNYGYTF TRBV2… TRBD0… TRBJ0… V28      D01      J01     
#>  2 TRB_CD8_949   CAIKMETPNGEQYF  TRBV1… NA     TRBJ0… V10      NA       J02     
#>  3 TRB_CD8_949   CAISESGVLNEKLFF TRBV1… NA     TRBJ0… V10      NA       J01     
#>  4 TRB_CD8_949   CASRDGQGSGNTIYF TRBV0… TRBD0… TRBJ0… V06      D01      J01     
#>  5 TRB_CD8_949   CASSEALPGMVPLHF TRBV0… TRBD0… TRBJ0… V02      D01      J01     
#>  6 TRB_CD8_949   CASSGGVAAFSSYN… TRBV0… TRBD0… TRBJ0… V05      D01      J01     
#>  7 TRB_CD8_949   CASSLAGDSQETQYF TRBV2… NA     TRBJ0… V28      NA       J02     
#>  8 TRB_CD8_949   CASSLQGREKLFF   TRBV2… TRBD0… TRBJ0… V27      D01      J01     
#>  9 TRB_CD8_949   CASSPAGAYYNEQFF NA     TRBD0… TRBJ0… NA       D02      J02     
#> 10 TRB_CD8_949   CASSPFDRGPDTEA… TRBV2… TRBD0… TRBJ0… V28      D01      J01     
#> 11 TRB_CD8_949   CASSPGTGTYGYTF  TRBV1… TRBD0… TRBJ0… V10      D01      J01     
#> 12 TRB_CD8_949   CASSPLDGLTNEQFF TRBV2… NA     TRBJ0… V28      NA       J02     
#> 13 TRB_CD8_949   CASSPPTGERDTQYF TRBV0… TRBD0… TRBJ0… V07      D02      J02     
#> 14 TRB_CD8_949   CASSPSRNTEAFF   TRBV0… TRBD0… TRBJ0… V04      D02      J01     
#> 15 TRB_CD8_949   CASSQDLGQAFSPL… TRBV0… TRBD0… TRBJ0… V04      D01      J01     
#> 16 TRB_CD8_949   CASSQDLMTVDSLF… TRBV0… TRBD0… TRBJ0… V04      D01      J02     
#> 17 TRB_CD8_949   CASSQDRTGQYGYTF TRBV0… TRBD0… TRBJ0… V04      D01      J01     
#> 18 TRB_CD8_949   CASSQDSSDTEAFF  TRBV0… NA     TRBJ0… V04      NA       J01     
#> 19 TRB_CD8_949   CASSQDWERLGEQFF TRBV1… TRBD0… TRBJ0… V14      D02      J02     
#> 20 TRB_CD8_949   CASSREGDQPQHF   TRBV0… TRBD0… TRBJ0… V05      D01      J01     
#> # ℹ 4 more variables: reading_frame <chr>, duplicate_count <int>,
#> #   duplicate_frequency <dbl>, seen <int>
# Track the top 10 clones from top.freq
LymphoSeq2::cloneTrack(
  study_table = amino_table,
  sequence_track = top_freq$junction_aa[1:10]
)
#> # A tibble: 75 × 12
#>    repertoire_id junction_aa     v_call d_call j_call v_family d_family j_family
#>    <chr>         <chr>           <chr>  <chr>  <chr>  <chr>    <chr>    <chr>   
#>  1 TRB_CD8_949   CAIKMETPNGEQYF  TRBV1… NA     TRBJ0… V10      NA       J02     
#>  2 TRB_CD8_949   CASRDGQGSGNTIYF TRBV0… TRBD0… TRBJ0… V06      D01      J01     
#>  3 TRB_CD8_949   CASSLQGREKLFF   TRBV2… TRBD0… TRBJ0… V27      D01      J01     
#>  4 TRB_CD8_949   CASSPFDRGPDTEA… TRBV2… TRBD0… TRBJ0… V28      D01      J01     
#>  5 TRB_CD8_949   CASSPGTGTYGYTF  TRBV1… TRBD0… TRBJ0… V10      D01      J01     
#>  6 TRB_CD8_949   CASSQDLGQAFSPL… TRBV0… TRBD0… TRBJ0… V04      D01      J01     
#>  7 TRB_CD8_949   CASSQDLMTVDSLF… TRBV0… TRBD0… TRBJ0… V04      D01      J02     
#>  8 TRB_CD8_949   CASSQDRTGQYGYTF TRBV0… TRBD0… TRBJ0… V04      D01      J01     
#>  9 TRB_CD8_949   CASSQDSSDTEAFF  TRBV0… NA     TRBJ0… V04      NA       J01     
#> 10 TRB_CD8_949   CASSREGDQPQHF   TRBV0… TRBD0… TRBJ0… V05      D01      J01     
#> # ℹ 65 more rows
#> # ℹ 4 more variables: reading_frame <chr>, duplicate_count <int>,
#> #   duplicate_frequency <dbl>, seen <int>
# Track clones mapping to the CD4 and CD8 samples while ignoring all others
LymphoSeq2::cloneTrack(
  study_table = amino_table,
  sample_list = c("TRB_CD4_949", "TRB_CD8_949")
)
#> # A tibble: 161 × 12
#>    repertoire_id junction_aa     v_call d_call j_call v_family d_family j_family
#>    <chr>         <chr>           <chr>  <chr>  <chr>  <chr>    <chr>    <chr>   
#>  1 TRB_CD4_949   CAISVGGSSPLHF   TRBV1… TRBD0… TRBJ0… V10      D02      J01     
#>  2 TRB_CD4_949   CASDGGFRNTIYF   TRBV1… TRBD0… TRBJ0… V19      D02      J01     
#>  3 TRB_CD4_949   CASGGLNTEAFF    NA     NA     TRBJ0… NA       NA       J01     
#>  4 TRB_CD4_949   CASGLVAGSTLGGE… TRBV1… TRBD0… TRBJ0… V12      D02      J02     
#>  5 TRB_CD4_949   CASGTGGETQYF    TRBV0… TRBD0… TRBJ0… V06      D02      J02     
#>  6 TRB_CD4_949   CASHSSGNTIYF    TRBV0… NA     TRBJ0… V06      NA       J01     
#>  7 TRB_CD4_949   CASKPPGQGGYGYTF TRBV0… TRBD0… TRBJ0… V06      D01      J01     
#>  8 TRB_CD4_949   CASMIDPSGNTIYF  TRBV0… NA     TRBJ0… V05      NA       J01     
#>  9 TRB_CD4_949   CASNARVDSPLHF   TRBV0… TRBD0… TRBJ0… V06      D01      J01     
#> 10 TRB_CD4_949   CASRLGESPLHF    NA     NA     TRBJ0… NA       NA       J01     
#> # ℹ 151 more rows
#> # ℹ 4 more variables: reading_frame <chr>, duplicate_count <int>,
#> #   duplicate_frequency <dbl>, seen <int>
# Track clones mapping to the CD4 and CD8 samples and track 2 specific 
#sequences
LymphoSeq2::cloneTrack(
  study_table = amino_table, sample_list = c("TRB_CD4_949", "TRB_CD8_949"),
  sequence_track = c("CASSPPTGERDTQYF", "CASSQDRTGQYGYTF")
)
#> # A tibble: 2 × 12
#>   repertoire_id junction_aa     v_call  d_call j_call v_family d_family j_family
#>   <chr>         <chr>           <chr>   <chr>  <chr>  <chr>    <chr>    <chr>   
#> 1 TRB_CD8_949   CASSPPTGERDTQYF TRBV07… TRBD0… TRBJ0… V07      D02      J02     
#> 2 TRB_CD8_949   CASSQDRTGQYGYTF TRBV04… TRBD0… TRBJ0… V04      D01      J01     
#> # ℹ 4 more variables: reading_frame <chr>, duplicate_count <int>,
#> #   duplicate_frequency <dbl>, seen <int>