Create a antigen specific T-cells database

This function takes as input path to the flat file versions of IEDB, McPAS-TCR, and VdjDB slim in that order and generates an RDA file of all T-cells and B-cells recorded to show antigenic specificity in these three databases. This is a script for internal use and will be run once every six months to update the databases. The flat files for this function can be downloaded at the following links:

IEDB: https://www.iedb.org/downloader.php?file_name=doc/receptor_full_v3.zip

McPAS-TCR: http://friedmanlab.weizmann.ac.il/McPAS-TCR/

VdjDB: https://github.com/antigenomics/vdjdb-db/releases Note: Here the vdjdb.slim.txt is used as the input Note: The order of the input matters, please provide the flat file paths in this order IEDB, McPAS-TCR, VdjDB.

Usage

prepDB(db_path = NULL)

Arguments

db_path

Path to IEDB, McPAS-TCR and VdjDB