Search for T cell receptor beta CDR3 amino acid sequences with known antigen specificity from IEDB, VdjDB and McPASDb

Search for published T cell receptor beta CDR3 amino acid sequences with known antigen specificity in a list of data frames.

Usage

searchDB(study_table, dbname = "all", chain = "trb")

Arguments

study_table

A tibble generated by the LymphoSeq2 functions readImmunoSeq(), productiveSeq(), searchPublished() or topSeqs(). "junction_aa", "duplicate_frequency", and "duplicate_count" are required columns.

dbname

A vector of database source to search for the antigenic specificity of a T-cell or B-cell in the dataset: "all", "IEDB", "McPAS-TCR", "VdjDB".

chain

The receptor chain type to search in the public databases: "tra", "trb", "light", "heavy".

Value

Returns the input table annotated with the any recorded antigenic specificity from the public databases.

Examples

file_path <- system.file("extdata", "TCRB_sequencing",
 package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
#> Dataset Analysis:
#>   Files: 10, Total: 0.00 GB, Largest: 0.0 MB
#>   Available memory: 11.6 GB
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- LymphoSeq2::productiveSeq(study_table = study_table,
 aggregate = "junction_aa")
top_seqs <- LymphoSeq2::topSeqs(productive_table = amino_table, top = 1)
LymphoSeq2::searchDB(study_table = top_seqs, dbname = "all", chain = "trb")
#> # A tibble: 10 × 26
#>    repertoire_id     junction_aa v_call d_call j_call v_family d_family j_family
#>    <chr>             <chr>       <chr>  <chr>  <chr>  <chr>    <chr>    <chr>   
#>  1 TRB_CD4_949       CASSESAGST… TRBV1… TRBD0… TRBJ0… V10      D02      J02     
#>  2 TRB_CD8_949       CASSPAGAYY… NA     TRBD0… TRBJ0… NA       D02      J02     
#>  3 TRB_CD8_CMV_369   CASSPPTGER… TRBV0… TRBD0… TRBJ0… V07      D02      J02     
#>  4 TRB_Unsorted_0    CASSPVSNEQ… TRBV2… TRBD0… TRBJ0… V28      D02      J02     
#>  5 TRB_Unsorted_1320 CASSPAGAYY… NA     TRBD0… TRBJ0… NA       D02      J02     
#>  6 TRB_Unsorted_1496 CASSPAGAYY… NA     TRBD0… TRBJ0… NA       D02      J02     
#>  7 TRB_Unsorted_32   CASSLQGREK… TRBV2… TRBD0… TRBJ0… V27      D01      J01     
#>  8 TRB_Unsorted_369  CASSQDWERL… TRBV1… NA     TRBJ0… V14      NA       J02     
#>  9 TRB_Unsorted_83   CASSLQGREK… TRBV2… TRBD0… TRBJ0… V27      D01      J01     
#> 10 TRB_Unsorted_949  CASSPAGAYY… NA     TRBD0… TRBJ0… NA       D02      J02     
#> # ℹ 18 more variables: reading_frame <chr>, duplicate_count <int>,
#> #   duplicate_frequency <dbl>, tra_cdr3_aa <chr>, gene <chr>, epitope <chr>,
#> #   pathology <chr>, antigen <chr>, tra_v_call <chr>, tra_j_call <chr>,
#> #   mhc_allele <chr>, reference <chr>, score <dbl>, cell_type <chr>,
#> #   source <chr>, trb_v_call <chr>, trb_j_call <chr>, Species <chr>