Similarity score

Calculates the similarity score of two samples.

Usage

similarityScore(sample_list)

Arguments

sample_list

A list of two tibble corresponding derived from the productiveSeq() function in LymphoSeq2. "duplicate_frequency", "junction_aa", and "repertoire_id" columns are necessary for the calculation of the Bhattacharyya coefficient.

Value

Returns the similarity score, a measure of the amount of overlap between two samples. The value ranges from 0 to 1 where 1 indicates the sequence frequencies are identical in the two samples and 0 indicates no shared frequencies.

See also