Top frequencies

Creates a table of the top productive amino acid sequences that have a specified minimum frequency threshold and reports the number of samples that the sequence appears in along with the minimum, maximum, and mean frequency across all samples. For T cell receptor beta sequences, the percent prevalence and antigen specificity of that sequence is also provided.

Usage

topFreq(productive_table, frequency = 0.1)

Arguments

productive_table

A tibble of productive amino acid sequences imported using the function LymphoSeq2 function productiveSeq() where the aggregate parameter was set to "junction_aa".

frequency

The minimum frequency that the sequence appears in any of the listed samples.

Value

A tibble of amino acid sequences and the number of samples that the sequence appears in along with the minimum, maximum, and mean frequency across all samples.

Details

For T cell receptor beta sequences, additionally reported is the \ Also provided is the antigen specificity of that sequence if known by comparing it to a database of previously reported sequences in the literature.

Examples

file_path <- system.file("extdata", "TCRB_sequencing",
 package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
#> Dataset Analysis:
#>   Files: 10, Total: 0.00 GB, Largest: 0.0 MB
#>   Available memory: 11.8 GB
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- LymphoSeq2::productiveSeq(study_table = study_table,
 aggregate = "junction_aa")
top_freq <- LymphoSeq2::topFreq(productive_table = amino_table,
 frequency = 0.1)