Clone tracking line plot

Creates line plot tracking amino acid frequencies across multiple samples

Usage

cloneTrack(study_table, sample_list = NULL, sequence_track = NULL)

Arguments

study_table

A tibble of productive amino acid sequences generated by LymphoSeq2 function productiveSeq where the aggregate parameter was set to "junction_aa"

sample_list

A character vector of one or more repertoire_ids to track. If set to NULL (default), all repertoire_ids in the sequence matrix will be tracked.

sequence_track

An optional character vector of one or more amino acid sequences to track. If set to NULL (default), will pull all junction_aa sequences from the sequence matrix.

Value

Returns a line plot showing the amino acid frequencies across multiple samples in the sequence matrix where each line represents one unique sequence.

Details

The plot is made using the package ggplot2 and can be reformatted using ggplot2 functions. See examples below.

See also

Examples

file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

atable <- productiveSeq(study_table = stable, aggregate = "junction_aa")

top_freq <- topFreq(atable, frequency = 0.001)

# Track clones without mapping or tracking specific sequences
cloneTrack(atable)
#> # A tibble: 7,533 × 12
#>    repert…¹ junct…² v_call d_call j_call v_fam…³ d_fam…⁴ j_fam…⁵ readi…⁶ dupli…⁷
#>    <chr>    <chr>   <chr>  <chr>  <chr>  <chr>   <chr>   <chr>   <chr>     <dbl>
#>  1 TRB_CD4… CAIKPG… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       3
#>  2 TRB_CD4… CAIRAG… TRBV1… TRBD2… TRBJ2… TRBV10  TRBD2   TRBJ2   in-fra…       3
#>  3 TRB_CD4… CAISDE… TRBV1… NA     TRBJ2… TRBV10  NA      TRBJ2   in-fra…       2
#>  4 TRB_CD4… CAISDL… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       2
#>  5 TRB_CD4… CAISDL… TRBV1… TRBD2… TRBJ1… TRBV10  TRBD2   TRBJ1   in-fra…       2
#>  6 TRB_CD4… CAISDQ… TRBV1… TRBD2… TRBJ1… TRBV10  TRBD2   TRBJ1   in-fra…       2
#>  7 TRB_CD4… CAISER… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       2
#>  8 TRB_CD4… CAISEW… TRBV1… NA     TRBJ2… TRBV10  NA      TRBJ2   in-fra…       4
#>  9 TRB_CD4… CAISGQ… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       2
#> 10 TRB_CD4… CAISLN… TRBV1… TRBD2… TRBJ2… TRBV10  TRBD2   TRBJ2   in-fra…       2
#> # … with 7,523 more rows, 2 more variables: duplicate_frequency <dbl>,
#> #   seen <int>, and abbreviated variable names ¹​repertoire_id, ²​junction_aa,
#> #   ³​v_family, ⁴​d_family, ⁵​j_family, ⁶​reading_frame, ⁷​duplicate_count

# Track top 20 clones mapping to the CD4 and CD8 samples
cloneTrack(atable, sample_list = c("TRB_CD4_949", "TRB_CD8_949"),
   sequence_track = top_freq$junction_aa[1:20]) 
#> # A tibble: 20 × 12
#>    repert…¹ junct…² v_call d_call j_call v_fam…³ d_fam…⁴ j_fam…⁵ readi…⁶ dupli…⁷
#>    <chr>    <chr>   <chr>  <chr>  <chr>  <chr>   <chr>   <chr>   <chr>     <dbl>
#>  1 TRB_CD4… CASSWE… TRBV2… TRBD1… TRBJ1… TRBV28  TRBD1   TRBJ1   in-fra…     174
#>  2 TRB_CD8… CAIKME… TRBV1… NA     TRBJ2… TRBV10  NA      TRBJ2   in-fra…     158
#>  3 TRB_CD8… CASRDG… TRBV6… TRBD1… TRBJ1… TRBV6   TRBD1   TRBJ1   in-fra…     303
#>  4 TRB_CD8… CASSEA… TRBV2… TRBD1… TRBJ1… TRBV2   TRBD1   TRBJ1   in-fra…      84
#>  5 TRB_CD8… CASSGG… TRBV5… TRBD1… TRBJ1… TRBV5   TRBD1   TRBJ1   in-fra…      62
#>  6 TRB_CD8… CASSLA… TRBV2… NA     TRBJ2… TRBV28  NA      TRBJ2   in-fra…     750
#>  7 TRB_CD8… CASSLQ… TRBV2… TRBD1… TRBJ1… TRBV27  TRBD1   TRBJ1   in-fra…    1957
#>  8 TRB_CD8… CASSLS… TRBV6… TRBD2… TRBJ2… TRBV6   TRBD2   TRBJ2   in-fra…     223
#>  9 TRB_CD8… CASSPF… TRBV2… TRBD1… TRBJ1… TRBV28  TRBD1   TRBJ1   in-fra…     295
#> 10 TRB_CD8… CASSPG… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…      67
#> 11 TRB_CD8… CASSPL… TRBV2… NA     TRBJ2… TRBV28  NA      TRBJ2   in-fra…     104
#> 12 TRB_CD8… CASSPP… TRBV7… TRBD2… TRBJ2… TRBV7   TRBD2   TRBJ2   in-fra…    1903
#> 13 TRB_CD8… CASSPS… TRBV4… TRBD2… TRBJ1… TRBV4   TRBD2   TRBJ1   in-fra…     106
#> 14 TRB_CD8… CASSQD… TRBV4… TRBD1… TRBJ1… TRBV4   TRBD1   TRBJ1   in-fra…     180
#> 15 TRB_CD8… CASSQD… TRBV4… TRBD1… TRBJ2… TRBV4   TRBD1   TRBJ2   in-fra…    2387
#> 16 TRB_CD8… CASSQD… TRBV4… TRBD1… TRBJ1… TRBV4   TRBD1   TRBJ1   in-fra…     223
#> 17 TRB_CD8… CASSQD… TRBV4… NA     TRBJ1… TRBV4   NA      TRBJ1   in-fra…      58
#> 18 TRB_CD8… CASSQD… TRBV1… TRBD2… TRBJ2… TRBV14  TRBD2   TRBJ2   in-fra…    1758
#> 19 TRB_CD8… CASSRE… TRBV5… TRBD1… TRBJ1… TRBV5   TRBD1   TRBJ1   in-fra…      67
#> 20 TRB_CD8… CASSSW… TRBV5… NA     TRBJ2… TRBV5   NA      TRBJ2   in-fra…      86
#> # … with 2 more variables: duplicate_frequency <dbl>, seen <int>, and
#> #   abbreviated variable names ¹​repertoire_id, ²​junction_aa, ³​v_family,
#> #   ⁴​d_family, ⁵​j_family, ⁶​reading_frame, ⁷​duplicate_count

# Track the top 10 clones from top.freq
cloneTrack(study_table = atable, sequence_track = top_freq$junction_aa[1:10]) 
#> # A tibble: 75 × 12
#>    repert…¹ junct…² v_call d_call j_call v_fam…³ d_fam…⁴ j_fam…⁵ readi…⁶ dupli…⁷
#>    <chr>    <chr>   <chr>  <chr>  <chr>  <chr>   <chr>   <chr>   <chr>     <dbl>
#>  1 TRB_CD8… CAIKME… TRBV1… NA     TRBJ2… TRBV10  NA      TRBJ2   in-fra…     158
#>  2 TRB_CD8… CASRDG… TRBV6… TRBD1… TRBJ1… TRBV6   TRBD1   TRBJ1   in-fra…     303
#>  3 TRB_CD8… CASSLQ… TRBV2… TRBD1… TRBJ1… TRBV27  TRBD1   TRBJ1   in-fra…    1957
#>  4 TRB_CD8… CASSPF… TRBV2… TRBD1… TRBJ1… TRBV28  TRBD1   TRBJ1   in-fra…     295
#>  5 TRB_CD8… CASSPG… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…      67
#>  6 TRB_CD8… CASSQD… TRBV4… TRBD1… TRBJ1… TRBV4   TRBD1   TRBJ1   in-fra…     180
#>  7 TRB_CD8… CASSQD… TRBV4… TRBD1… TRBJ2… TRBV4   TRBD1   TRBJ2   in-fra…    2387
#>  8 TRB_CD8… CASSQD… TRBV4… TRBD1… TRBJ1… TRBV4   TRBD1   TRBJ1   in-fra…     223
#>  9 TRB_CD8… CASSQD… TRBV4… NA     TRBJ1… TRBV4   NA      TRBJ1   in-fra…      58
#> 10 TRB_CD8… CASSRE… TRBV5… TRBD1… TRBJ1… TRBV5   TRBD1   TRBJ1   in-fra…      67
#> # … with 65 more rows, 2 more variables: duplicate_frequency <dbl>, seen <int>,
#> #   and abbreviated variable names ¹​repertoire_id, ²​junction_aa, ³​v_family,
#> #   ⁴​d_family, ⁵​j_family, ⁶​reading_frame, ⁷​duplicate_count

# Track clones mapping to the CD4 and CD8 samples while ignoring all others
cloneTrack(study_table = atable, sample_list = c("TRB_CD4_949", "TRB_CD8_949"))
#> # A tibble: 1,607 × 12
#>    repert…¹ junct…² v_call d_call j_call v_fam…³ d_fam…⁴ j_fam…⁵ readi…⁶ dupli…⁷
#>    <chr>    <chr>   <chr>  <chr>  <chr>  <chr>   <chr>   <chr>   <chr>     <dbl>
#>  1 TRB_CD4… CAIKPG… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       3
#>  2 TRB_CD4… CAIRAG… TRBV1… TRBD2… TRBJ2… TRBV10  TRBD2   TRBJ2   in-fra…       3
#>  3 TRB_CD4… CAISDE… TRBV1… NA     TRBJ2… TRBV10  NA      TRBJ2   in-fra…       2
#>  4 TRB_CD4… CAISDL… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       2
#>  5 TRB_CD4… CAISDL… TRBV1… TRBD2… TRBJ1… TRBV10  TRBD2   TRBJ1   in-fra…       2
#>  6 TRB_CD4… CAISDQ… TRBV1… TRBD2… TRBJ1… TRBV10  TRBD2   TRBJ1   in-fra…       2
#>  7 TRB_CD4… CAISER… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       2
#>  8 TRB_CD4… CAISEW… TRBV1… NA     TRBJ2… TRBV10  NA      TRBJ2   in-fra…       4
#>  9 TRB_CD4… CAISGQ… TRBV1… TRBD1… TRBJ1… TRBV10  TRBD1   TRBJ1   in-fra…       2
#> 10 TRB_CD4… CAISLN… TRBV1… TRBD2… TRBJ2… TRBV10  TRBD2   TRBJ2   in-fra…       2
#> # … with 1,597 more rows, 2 more variables: duplicate_frequency <dbl>,
#> #   seen <int>, and abbreviated variable names ¹​repertoire_id, ²​junction_aa,
#> #   ³​v_family, ⁴​d_family, ⁵​j_family, ⁶​reading_frame, ⁷​duplicate_count

# Track clones mapping to the CD4 and CD8 samples and track 2 specific sequences
cloneTrack(study_table = atable, sample_list = c("TRB_CD4_949", "TRB_CD8_949"),
   sequence_track = c("CASSPPTGERDTQYF", "CASSQDRTGQYGYTF"))
#> # A tibble: 2 × 12
#>   reperto…¹ junct…² v_call d_call j_call v_fam…³ d_fam…⁴ j_fam…⁵ readi…⁶ dupli…⁷
#>   <chr>     <chr>   <chr>  <chr>  <chr>  <chr>   <chr>   <chr>   <chr>     <dbl>
#> 1 TRB_CD8_… CASSPP… TRBV7… TRBD2… TRBJ2… TRBV7   TRBD2   TRBJ2   in-fra…    1903
#> 2 TRB_CD8_… CASSQD… TRBV4… TRBD1… TRBJ1… TRBV4   TRBD1   TRBJ1   in-fra…     223
#> # … with 2 more variables: duplicate_frequency <dbl>, seen <int>, and
#> #   abbreviated variable names ¹​repertoire_id, ²​junction_aa, ³​v_family,
#> #   ⁴​d_family, ⁵​j_family, ⁶​reading_frame, ⁷​duplicate_count